pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.0"
ipython = "^9.0.2"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
stripe = "^12.0.0"
drf-yasg = "^1.21.10"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
